Add step for every part of the registration form et create a layout repairer registration layout
